Context Interface

Building a Context

CompetingClocks.SamplingContext — Method

SamplingContext(::Type{K}, ::Type{T}, rng::AbstractRNG;
    step_likelihood=false,
    path_likelihood=false,
    debug=false,
    recording=false,
    common_random=false,
    method=nothing,
    start_time::T,
    likelihood_cnt::Int
    )

The SamplingContext is responsible for doing a perfect forwarding to multiple components that implement the desired features in a sampler.

It uses a Context pattern where each type parameter is either present or Nothing, and the compiler knows to optimize-away the Nothing. We keep the internal logic simple. If something is present, call it.

memory

delayed transitions

hierarchical samplers

K and T are the clock type and time type.

step_likelihood - whether you will call steploglikelihood before each fire!

path_likelihood - whether you will call pathloglikelihood at the end of a simulation run.

debug - Print log messages at the debug level. Enabling this stores every enabling and disabling event so don't leave it on.

recording - Store every enable and disable for later examination.

common_random - Use common random numbers during sampling.

method - If you want a single, particular sampler, put its SamplerSpec here. It will create a group called :all that has this sampling method.

start_time - Sometimes a simulation shouldn't start at zero.

likelihood_cnt - The number of likelihoods to compute, corresponds to number of distributions in enable! calls. This turns on path_likelihood.

Example

builder = SamplerBuilder(Tuple,Float64)
add_group!(builder, :sparky => (x,d) -> x[1] == :recover, method=NextReaction())
add_group!(builder, :forthright=>(x,d) -> x[1] == :infect)
context = SamplingContext(builder, rng)

CompetingClocks.SamplingContext — Method

SamplingContext(builder::SamplerBuilder, rng)

Uses the SamplerBuilder to make a SamplingContext.

CompetingClocks.SamplerBuilder — Type

SamplerBuilder(::Type{K}, ::Type{T};
    step_likelihood=false,
    path_likelihood=false,
    debug=false,
    recording=false,
    common_random=false,
    method=nothing,
    start_time::T,
    likelihood_cnt::Int
    )

A SamplerBuilder is responsible for recording a user's requirements and building an initial sampler.

K and T are the clock type and time type.

step_likelihood - whether you will call steploglikelihood before each fire!

path_likelihood - whether you will call pathloglikelihood at the end of a simulation run.

debug - Print log messages at the debug level. Enabling this stores every enabling and disabling event so don't leave it on.

recording - Store every enable and disable for later examination.

common_random - Use common random numbers during sampling.

method - If you want a single, particular sampler, put its SamplerSpec here. It will create a group called :all that has this sampling method.

start_time - Sometimes a simulation shouldn't start at zero.

likelihood_cnt - The number of likelihoods to compute, corresponds to number of distributions in enable! calls. This turns on path_likelihood.

Example

builder = SamplerBuilder(Tuple,Float64)
add_group!(builder, :sparky => (x,d) -> x[1] == :recover, method=NextReaction())
add_group!(builder, :forthright=>(x,d) -> x[1] == :infect)
context = SamplingContext(builder, rng)

CompetingClocks.add_group! — Function

The selector defines the group of clocks that go to this sampler using an inclusion rule, so it's a function from a clock key and distribution to a Bool.

Choosing Sampler Methods

CompetingClocks.NextReactionMethod — Type

NextReactionMethod()

Uses Anderson's Modified Next Reaction method for distributions in an Exponential class (Exponential, Weibull, Erlang) and the Next Reaction method for other distributions. Yes, these two methods are mixed into one sampler because Julia's holy traits pattern makes it efficient to use the fastest sampler on a distribution-by-distribution basis. Because it reuses draws, this is the best choice if you want to do variance reduction with Common Random Numbers.

See CombinedNextReaction.

CompetingClocks.DirectMethod — Type

DirectMethod()
DirectMethod(memory::Symbol, search::Symbol)
DirectMethod(:keep, :scan)
DirectMethod(:keep, :tree)
DirectMethod(:remove, :scan)
DirectMethod(:remove, :tree)

Use this to specify any Direct method for Exponential distributions. Defaults to memory=:remove so it limits memory growth over time but memory=:keep will be faster if the space of clock keys is limited. Defaults to search=:tree for best performance for many enabled clocks but search=:scan is faster for small numbers of clocks. The different kinds of methods, like "Optimized Direct Methods" amount to using different computer science techniques for scanning sums of hazard rates, and that's what the search algorithm lets you choose.

See DirectCall.

CompetingClocks.FirstReactionMethod — Type

FirstReactionMethod()

The classic sampler that draws every clock at every time step. Very fast for very small numbers of enabled clocks and returns a new next() every time it is called which helps when resampling paths. Other samplers return the same value every time you call next() unless you jitter! them, which is expensive.

See FirstReaction.

CompetingClocks.FirstToFireMethod — Type

FirstToFireMethod()

The simplest and fastest sampler. When you enable!() a clock, this draws the firing time and saves it in a queue.

See FirstToFire.

CompetingClocks.PartialPropensityMethod — Type

PartialPropensityMethod()

Exact, continuous-time sampler using composition-rejection over groups. Only for exponential distributions. This variant of partial-propensity composition-rejection implements the sampler but not the reaction network. The reaction-network can be implemented outside of the sampler.

See PSSACR.

CompetingClocks.PetriMethod — Type

PetriMethod()
PetriMethod(dt)

Samples by picking at random ignoring distributions. Good for testing rare cases in simulations. Increments time dt=1.0 by default. It's called "Petri" because a Petri net model always chooses the next event at random.

See Petri.

CompetingClocks.RejectionMethod — Type

RejectionMethod(bound_factor=1.05)

A rejection-based algorithm. Only for exponential distributions, this may be the fastest for large simulations. The bound_factor>= 1.0 controls the default upper bounds. Set to 1.0 for no rejections, which reduces this to the direct method.

See RSSA.

Using a Context

CompetingClocks.enable! — Method

enable!(sampler, clock, distribution, relative_te)

Tell the sampler to start a clock with a time shift to the left or right.

sampler::SSA{KeyType,TimeType} - The sampler to tell.

clock::KeyType - The ID of the clock. Can be a string, integer, tuple, etc.

distribution::Distributions.UnivariateDistribution

relative_te::TimeType - The zero-point of the distribution relative to the current simulation time. A -0.1 shifts the distribution to the left. A 0.1 says the clock cannot fire until after 0.1 units of time have passed.

CompetingClocks.enable! — Method

enable!(sampler, clock, distribution)

Tell the sampler to start a clock.

sampler::SSA{KeyType,TimeType} - The sampler to tell.

clock::KeyType - The ID of the clock. Can be a string, integer, tuple, etc.

distribution::Distributions.UnivariateDistribution

CompetingClocks.enable! — Method

enable!(sampler, clock, distribution::Vector, relative_te)

Vectorized version of enable! for multiple-distributions in likelihood. This will sample times from the first distribution in the vector unless sample_from_distribution! is set. All distributions are used to calculate a path likelihood for importance sampling.

sampler::SSA{KeyType,TimeType} - The sampler to tell.

clock::KeyType - The ID of the clock. Can be a string, integer, tuple, etc.

distribution::Vector{Distributions.UnivariateDistribution}

relative_te::TimeType - The zero-point of the distribution relative to the current simulation time. A -0.1 shifts the distribution to the left. A 0.1 says the clock cannot fire until after 0.1 units of time have passed.

CompetingClocks.enable! — Method

enable!(sampler, clock, distribution::Vector)

Vectorized version of enable! for multiple-distributions in likelihood. This will sample times from the first distribution in the vector unless sample_from_distribution! is set. All distributions are used to calculate a path likelihood for importance sampling.

sampler::SSA{KeyType,TimeType} - The sampler to tell.

clock::KeyType - The ID of the clock. Can be a string, integer, tuple, etc.

distribution::Vector{Distributions.UnivariateDistribution}

This version enables the clock with no time-shifting.

CompetingClocks.disable! — Method

disable!(sampler, clock)

Tell the sampler to forget a clock. The clock will be disabled at the time set by the last call to fire!.

CompetingClocks.next — Method

next(ctx::SamplingContext)

Return (when, clock-key) for the next event. This does NOT fire the event. You can check if when is past the end time of the simulation and choose not to fire that event. For samplers like FirstReactionMethod, you could call this many times to generate many possible events.

CompetingClocks.fire! — Method

fire!(ctx::SamplingContext, clock, when)

Decide the next clock key to fire and the time at which it fires. This sets the internal time of the simulation to the new time when. It also disables the given clock key.

CompetingClocks.reset! — Method

reset!(sampler)

After a sampler is used for a simulation run, it has internal state. This function resets that internal state to the initial value in preparation for another sample run.

reset!(sampling)

Clears all clock values in a SamplingContext at the top of a loop over simulations. Call this instead of allocating a new sampler for each iteration.

CompetingClocks.copy_clocks! — Method

copy_clocks!(dst::SamplingContext, src::SamplingContext)

Used for splitting a simulation near a rare event. You keep a vector of samplers and once your current sampler reaches a simulation state, copy its state into the vector of samplers and pick up where it left off. The destination SamplingContext has its own random number generator which will not be overwritten. The destination sampler will be jittered so that it will return different next() samples from the same set of enabled clocks.

CompetingClocks.split! — Method

split!(dst::AbstractVector{S<:SamplingContext}, src::SamplingContext{K,T})

If the src sampler has gotten to a good spot in the simulation, split it into multiple copies in the destination vector. Update the split_weight member of each destination copy by 1/length(dst). The destination can be a view of a vector that was created with clone().

Base.Libc.time — Method

time(sampling)

Get the current simulation time. Simulation times are incremented by each call to fire!.

CompetingClocks.sample_from_distribution! — Method

sample_from_distribution!(sampling, index)

If you created a SamplerBuilder with likelihood_cnt > 1, then you are going to call enable! for some or all clocks using a vector of distributions. In that case, the SamplingContext will default to using the first distribution to decide which event comes next. That way you get a sample from distribution 1 and likelihoods from distributions (1, 2, and so on). For doing mixed importance sampling, you will sample from several distributions and calculate a weighted average. This functions lets you choose which of the enabled distribution vector to sample.

CompetingClocks.clone — Method

clone(sampling, rng)

Given a SamplingContext, make a copy as though you had called the constructor again. This is useful for creating a vector of SamplingContext for multi-threading or for splitting paths. This clone will clone every part of the object. Note that the random number generator is copied, and you will want that to be unique for each clone.

You need to give each sampler its own random number generator, rng. If you were making parallel RNGs with the Random123 package, it might look like:

julia master_seed = (0xd897a239, 0x77ff9238) rng = Philox4x((0, 0, 0, 0), master_seed) Key = Int64 sampler = SamplingContext(SamplerBuilder(Key,Float64; path_likelihood=true), rng) observation_weight = zeros(Float64, particle_cnt) total_weight = zeros(Float64, particle_cnt) samplers = Vector{typeof(sampler)}(undef, particle_cnt) for init_idx in 1:particle_cnt rng = Philox4x((0, 0, 0, init_idx), master_seed) samplers[init_idx] = clone(sampler, rng) init!(state[init_idx], samplers[init_idx]) # For some initialization of state. end`

CompetingClocks.freeze_crn! — Method

freeze_crn!(ctx::SamplingContext)

After running the simulation to collect random draws, call this function to stop collection of new draws and solely replay the draws that were collected, using only fresh draws for clocks that weren't seen before.

CompetingClocks.reset_crn! — Method

reset_crn!(ctx::SamplingContext)

This resets a sampler including erasing its stored common random numbers. Using a simple reset!(sampler) won't erase the saved CRN draws.

CompetingClocks.enabled — Method

enabled(sampling)

Returns a set of enabled clock keys. This will aggregate enabled clock keys across hierarchical samplers.

CompetingClocks.enabled_history — Method

enabled_history(ctx::SamplingContext)

Returns a Vector{EnablingEntry{K,T}} that has every time a clock was enabled.

See CompetingClocks.EnablingEntry.

CompetingClocks.disabled_history — Method

disabled_history(ctx::SamplingContext)

Returns a Vector{DisablingEntry{K,T}} that has every time a clock was enabled.

See CompetingClocks.DisablingEntry.

CompetingClocks.EnablingEntry — Type

EnablingEntry{K,T}(clock::K, distribution, te::T, when::T)

Records that clock is enabled at time when with an enabling time te (in absolute time) that sets the zero of the distribution.

Fields

clock::K: The key for the event/clock/transition.

distribution::Distributions.UnivariateDistribution, a distribution of clock firing times.

te::T - An absolute time to use as the zero-time for the distribution. Usually the same as when.

when::T - The time this clock was enabled.

CompetingClocks.DisablingEntry — Type

DisablingEntry{K,T}(clock::K, when::T)

Records that clock is disabled at time when.

Fields

clock::K: The key for the event/clock/transition.

when::T - The time this clock was enabled.

Base.length — Method

length(sampling)

The total number of enabled clocks.

CompetingClocks.isenabled — Method

isenabled(sampling, clock)

Boolean for whether this clock key is currently enabled, checked across hierarchical samplers.

Base.keytype — Method

keytype(sampler)

Return the type of clock keys.

keytype(sampling)

The type for the clock key. If you can use a concrete type, that helps performance.

CompetingClocks.timetype — Method

timetype(sampler)

Return the type of clock times.

timetype(sampling)

The type for the time, usually a Float64.

CompetingClocks.steploglikelihood — Method

steploglikelihood(sampling, when, which)

If the next event had clock key which and happened at time when, what would be the log-likelihood of that event conditioned on the last event? This calculates relative to the last call to fire!(). If you were to integrate this value over all enabled events, from the last firing time to Inf with the QuadGK package, it would sum to one.

CompetingClocks.pathloglikelihood — Method

pathloglikelihood(sampling, endtime)

Calculates the log-likelihood across all events since the start of the simulation and up to the given end time. We include the end time because some statistics are normalized not to the last event but to a specific time, so this includes the probability that no event fired before the given end time.

This value can be used to weight paths for importance sampling.